fix(design-system): refactor componentEntries logic and update imports
